+++ docker/docker_start.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
-node bin/cli database migrate filaments
+# exit on any error
+set -e
 
+# wait for services (db, cache) to be ready
+node bin/wait-services
+# run database migration
+node bin/cli database migrate filaments
+# start the application
 yarn start